requirements.txt
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
aiohttp>=3.6.2
aiohttp-cors>=0.7.0